small state refac
