+++ src/eth/storage/temporary/inmemory/transaction.rs
@@ -5,12 +5,12 @@ use parking_lot::RwLockUpgradableReadGuard;
 #[cfg(not(feature = "dev"))]
 use parking_lot::RwLockWriteGuard;
 
-use crate::eth::executor::Complete;
 #[cfg(feature = "dev")]
 use crate::eth::executor::CompleteValue;
 use crate::eth::executor::State;
 use crate::eth::executor::TransactionExecution;
 use crate::eth::executor::TransactionExecutionInput;
+use crate::eth::executor::types::state::Complete;
 use crate::eth::storage::StorageError;
 use crate::eth::storage::TxCount;
 use crate::eth::storage::temporary::inmemory::InMemoryTemporaryStorageState;
